Specify which address Docker should bind to
While deploying to Digital Ocean I got an error because Docker could not determine which address to bind to. eth0 had 2 ipv4 addresses (one internal and one external). This specifies `advertise_addr` to use the same IP address that ansible is using to access the remote machine.

However, maybe we need something else in order to support ipv6 as well?